simplification and improvement of the Phone entity with update of the fixtures linked to it
